backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
sqlalchemy==2.0.27
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.0
python-multipart==0.0.9
websockets==12.0
scikit-learn==1.4.1.post1
pandas==2.2.1
numpy==1.26.4
pytesseract==0.3.10
Pillow==10.2.0
reportlab==4.1.0
matplotlib==3.8.3
python-dotenv==1.0.1
joblib
reportlab
pydantic[email]
python-socketio
opencv-python-headless
pdfplumber
psycopg2-binary  # PostgreSQL driver
gunicorn         # Production-grade web server